Let equals match const and non-const objects
- Non-const and const variants of objects will now return true when using equals, because they can represent the same object, just with a different access policy.

+++ bigdoors-core/src/test/java/nl/pim16aap2/bigdoors/util/CuboidTest.java
@@ -17,7 +17,7 @@ class CuboidTest
         final @NotNull Vector3Di max = new Vector3Di(10, -10, 10);
         final @NotNull Cuboid cuboid = new Cuboid(min, max);
 
-        Assert.assertEquals(11 * 11 * 11, cuboid.getVolume().intValue());
+        Assert.assertEquals(11 * 11 * 11, cuboid.getVolume());
     }
 
     @Test
